V1.0
Basic Gameplay has been added. The ability to have a mini-map in the corner that works and accurately depict the player's location is fully implemented, with the ability to move through rooms conditionally. Branching dialogue has been added as well, with choices that result in key variables being changed with conditional requirements. Dialogue has been partially implemented, such that it can display one string of dialogue for any given action.