v1.1
Major Update to Dialogue: We are now able to chain multiple strings together for Dialogue. If we are entering a room for the first time, it will display all of the  In theory, we're able to do this for events as well though at the moment we're using multiple strings for branching dialogue as a way to do "Yes/No" options, as well as "Can/Can't". In theory, we will now be able to sit down and prepare events and write the actual dialogue for the game, and prepare a title screen in the future.