v1.4
Larger Update - Added the ability for text to dynamically linebreak once it reaches a certain length, with that length being able to be changed at will. Furthermore, it will also check to see if a word would be broken and adjusting where the linebreak occurs to compensate. With this, a larger overhaul of previous text is necessary and a restructuring may be helpful, but it will cause adding text to be much easier overall without needing to manually add it.