v1.5
I have begun to unspaghetti the code and make it much more reusable. I have also continued working on making the typewriter text work properly, so now it makes a sound and also displays each character and writes it out in real time. Hopefully the next update will continue to unspaghetti it, and add functionality for paragraphs