v1.7
v1.7

We have fully reimplemented the branching dialogues, and have made it work dynamically. Prior, we could only have two lines of dialogue for any branching option, with one being 'new' and one being 'revisiting'. However, now we can have any amount of 'new' dialogue. We have also went around and cleaned up our code and fixed a variety of bugs that caused sound or dialogue to be incomplete.

We will need to fix the events dialogue, but once we do so the game will be fully functional and we can focus on writing the story properly without concern.