v1.8
v.18 - Events have been implemented. We're now at the stage where all that's left is to write the majority of the story, clean up the code, and add events as it comes up. If possible, I may add a title screen with adjustable options, and I'd like to review the text program to ensure that it's displaying properly, and possibly even rework it so it can display paragraphs. For now however, it's finally working!