v1.9
v1.9 - Added functionality for Room 5. We are in early versions of it, so I still need to fully implement the Journal 1 text for its trigger, but otherwise I have implemented the way to switch between multiple versions of the same room. In the next update, I should hopefully finish Room 5! I probably shouldn't have started by doing v1.x, since it's still not a complete game, but it's a little late to fix.