v1.10
v1.10 - Added Room 4 functionality, and added logic for the Doors of Room 5 to properly "remember" if you've tried either door 4 or door 6. I'll need to condense and make it cleaner, but it works for now. Room 5 should hopefully have been the hardest room in the game to deal with, considering how you have to move through it. Only Room 9, 11 and 12 should pose issues in this manner as other rooms are self contained. Flipped the map so it accurately depicted how the map should look.