v1.11
v1.11 - Fully reworked the previous patches additions. Previous methods to accommodate for 4+ options in a room have been removed due to poor potential for growth on anything past 4 options. Instead, it has been reworked such that if there are more than 3 options, you can scroll through the options to view them all, with a visual indicator to let you know if you can scroll, as well as if there's anything left to check still. This should heavily condense the program and reduce headache for any future changes, and make Events easier to read. Some comments have been removed, and will be readded in a future patch.