v1.12
Major rework for UI - Text will display incorrectly at this time due to the issues, but will be solved by rewriting the text and properly understanding new limits. Minor Room 6 work has been done, and the groundwork for "preview" rooms are in place, i.e: if a player can see a door of a room they haven't entered yet, they can see it on the map to understand the layout of the land better. Only the preview for room 4 has been done for now, but other rooms can easily have this added.