v1.13
v1.13 - We have added the functionality for two new abnormalities. In this case: When we have less than three choices, the program will now display this accurately without showing a "nil" choice. With some minor work in the future, we may also be able to add functionality so we can always save the cursor's position instead of needing to reset it after checking every new "subroom". We have also finally created the ability for artificial linebreaks within the program, and for them to display appropriately without fear of shortened lines or incomplete words via a set fake character.