1.15
Major Code Rework - Decluttering of many antiquated variables, and splitting the code into functions for better reading and ease of adjustment and reuse as needed. Program is getting ready for a title screen. Next update should have the Room 4 updates, and hopefully a title screen with adjustable options for text color, border color, and sound.