v1.18
v1.18 - Updated Inventory menu, functionality has increased to it's fullest potential. All we are missing is the ability for the save code to be generated properly. Room 9 is now fully implemented, as are the proper events for it. As expected and feared, we are still running dangerously close to the character limit by our predictions: We have less than 24k characters to complete everything. Future updates will look at Rooms 10-12 as we prepare to reach the "end game" of the project, while also ensuring that each word we type for the story is worth putting in.