1.19
1.19 - Revamp of the Inventory Mechanic, it is now Dynamic and can support any positive number of entries and adjust accordingly. Of course, after a certain amount, it may start to display incorrectly if the goal is to have more than 12 entries, but this should not be an issue with the scope of our program, as we should not have more than 8 entries. In the event that we need 13+ entries, we can update the code to work similar to our Choice system in our dialogue.

Minor revamps to Room 9 and 12, to better incorporate the ability to see our first instance of our monster while experimenting with the ability to take away choice and override rooms. This will be used in the future for Rooms 11, 12, and 9 again during the "Escape" sequence.

Currently: The gameplan with the story is that the player must do one of the following:

Use a metal pole to try and wedge the door open, and enter, and the door snaps shut behind them
Use a metal pole to make a loud sound, and the monster comes, investigates, and breaks down the door
Uses a half empty bottle of herbicide to deal with the vines in the door.

There are issues at play: It is preferred that we encourage the player to look at the monster in the hallway before moving to Room 11, thus one solution is that we use Option 2 where the metal pole just doesn't work, and the player is forced to draw the attention of the monster to open the door, but it feels unrealistic for the player character to do this and not just run out the exit door. Using Option 1 is idea in ensuring that it makes a loud enough sound that storywise it draws attention, but otherwise feels a bit odd in that the player shouldn't be strong enough to jam the door open or wedge it. Finally, using the herbicide is a third option, though we intend to use it later on, so having two different items just for herbicide feels odd, even if we flavor that after using the first one, it runs out. This will be solved by the next update.