1.20.6
1.20.6 - Updates to (almost) all rooms, we're focusing on a new design philosophy to have the "Look through Room" option available as the first prompt that the player can choose, which involve minimal swapping of choices. This should normalize the experience between rooms. As of now: We have elected not to have the "Hallway" of Room 9 be a room you can look through for budgets sake, but we can add it if we have the space in the end.

We have continued looking through and revamping the rooms, while also trimming fluff code. As it stands, we're getting close to being done with this draft of writing, and we're much happier about how this has gone. We've also started encouraging more reusability with the "scanning" dialogue and making it more efficient with our code.

We've also removed more story elements that didn't impact the story, and replaced them with proper story beats and ideals. We should hopefully be close to the end of our revamps by next update, and can continue with actual updates.