1.21.5
1.21.5 - Room 10 is being prepared. The Alpha Main Room Dialogue and the Initial Choice 4 Dialogue has been added, and our monster is now present, even with a unique sprite on the map, albeit only four pixels. Once again: We have reached an issue where we're desperately low in space. As it stands, we only have 11k characters to finish Room 10, and then backtrack to Room 11, 12, 9, and finally do ending "run away", or go to rooms 7 and 8 for the first time.

The budgets low, and once again it's seeming to be rather clear that cutting more dialogue and content may be necessary. Over the next few days, review over what is necessary for the story, and what isn't will happen. In a perfect world: We'll only need another 2-3k characters for Room 10, and can get by on 1k for Room 11 and 12 each, though it's likely we may higher on Room 11 due to an optional key and keycard being present. Room 9 in comparison should only need 1k, and the Neutral Ending should only take 1k.

I'd ***love*** to have Room 8 have content, but it's likely that we'll need to cut Room 8 for now and make it a "you run through this room" type of event. Room 7 should take about 3-5k characters, in theory. It's an ending, not a room in reality as once you enter, you only have a few choices. If we can continue to cut down and trim, it should make it possible to finish the story. It's highly regrettable that we must continue to trim on the story, but as it stands while I would like to preserve my style of writing and explanation, the game can't be completed without some sacrifice. I'll have to budget better for the future, though ideally a character limit shouldn't be as relevant again.