1.22.1
1.22.1 - Ending "D" is now written. As it stands, we're still experimenting with what to do going forward to ensure that wasted space is no longer a concern. Some items can be moved around to try and ensure certain rooms aren't overly "stacked" with interactions, and give an even spread of information. As it stands, we're still desperately low in space, but it may be best to finish writing the story first, and then deal with consolidation at the end to avoid an ouroboros of endless consolidation phases.

This should also work for the goal of knowing exactly how we need to save as well. Ideally: A binary system can be used understand what details to share, but we'll need to dedicate time to that only after the story is nearing completion, as we don't wish to revamp the system as we're adding information.