1.22.2
1.22.2 - The "Good" Ending is complete, in writing at least, and now we can officially claim that the "alpha" version of Room 7 is now complete, as are all the Endings. Speaking bluntly, I'm glad to see this as it is. However: We once again find ourselves in dire straits, where to finish everything else, we only have 4k characters to work with. Allow me to display what still must be done for this program to be considered "complete" and "playable".

- The Save/Load System needs to be addressed.
- Room 12 must be completed, it is at 30% completion, roughly.
- Room 11 Variant B is untouched.
- Rooms 10 and 9 must have at least simple "going through" dialogue.
- Room 8 is untouched, outside of the simplest coding required for Room 7.

We must do this within 4k characters. Thus: Once again we're at a stage where we must cut content in order to make room. It goes without saying that while we can certainly condense the actual logic of the program further and save a few hundred characters, that isn't enough to solve our issue. As noted in 1.22.1, we can't exactly start begin Revisions to save a few thousand characters every room. We're continuing with the original plan to write everything there is to write, and then after we're far and away above our "limit", we would then condense down to manageable territory.

With that said, this update may seem "meaningless", as it goes against the entire premise that we have just put onto the screen. However, I do value this as an important update due to the finality of the endings being completed, and wishing to cover any issues in the event of system problems. We're also working with a revamp of a certain "plot point".

"Evacurine", i.e the former Item 3, was an item that could be taken if you went out of your way to grab it in Room 5. The intended goal and method was to use it in the "good ending". In this instance, the player would give Eva the Evacurine, and she would react positively. Within the original variant of "A Forgotten Journal", this plot point did not exist, though we can hardly use that as an excuse considering that many additions of here did not exist in the original. As an example, the chains binding Eva in Room 7 did not properly exist. However, when it comes to the subject of cutting content, it's not hard to see why this is up for removal. The entire ordeal takes a thousand characters, and the concept of doing a good deed to help Eva is already addressed by removing her chains and setting her free, while also still protecting her.

I do value Evacurine for Eva's character, as it's an important aspect of her character and story, but we're at an impasse for how to actually implement it anyhow, and we're desperate for either finding space, or reworking old content into another area. As it stands, we have moved it from Room 5, and may either put it in Room 10, or alternatively in Room 8. The goal is simple, to allow it as a use to get past Room 8 and gain access to Room 7. In theory, it could be used to "feed" the plants that are blocking Room 7's door, though a method may still need to be used to make it feel natural.

With this tangent aside, I have removed the plotpoint from Room 5, as it takes a 1k characters that at this point were useless. We'll likely be forced to remove more plotpoints, though I would wish to avoid removing too many "nothing" options, as I do want this to be a CYOA. If I only offer options that give progression and never an option that exists for the sake of story or fun, then it would go against my wishes. This rework has the added benefit of making inventory management *much* nicer. Really, it's a stain on this program that I had an inventory system that worked in effective binaries of 0 or 1, and had an item that started with -1. This should make managing the Inventory for saves much easier, as with a binary system.

Of course, at this stage it may be easier to simply hard code certain saves, but I would love to avoid hard coding as much as reasonably possible. Hammers are effective tools for flattening problems, but one wouldn't want to take a hammer to their keyboard simply because a single key has something stuck in it. Thus, we may be able to benefit from needing two binary checks, for explored rooms and for inventory. Of course, a variety of complications exist, such as various "checks" that the story has to update or create options, but those are being weighed as well.

Notably: To bring this update to a close, I do not intend to post another update until the Story aspect of this program is complete. Once I have fully written the story and condensed it, we will finally be ready to reach the ending stages of this program. Should we be lucky, I would hope to complete this program in its entirety by 7/16